backend/requirements.txt
# Reproducible environment for backend (Windows-friendly)
# Pin these to the versions we validated together
fastapi==0.121.1
uvicorn==0.38.0
python-dotenv==1.2.1

# LangChain 1.x family (modern LCEL APIs)
langchain==1.0.5
langchain-core==1.0.4
langchain-community==0.4.1
langchain-huggingface==1.0.1
langchain-groq==1.0.0
langchain-text-splitters==1.0.0

# Vector store
faiss-cpu==1.12.0

# Embeddings / HF stack
transformers==4.57.1
huggingface-hub==0.36.0
sentence-transformers==5.1.2

# Speed / parsing conveniences
orjson==3.11.4
ormsgpack==1.12.0

# Optional: for file loading if you extend ingest
pypdf==6.2.0
filetype==1.2.0
